--- homework05/src/test/java/me/vladislav/homework/HomeworkApplicationTests.java
@@ -1,15 +1,0 @@
-package me.vladislav.homework;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class HomeworkApplicationTests {
-
-  @Test
-  void contextLoads() {
-  }
-
-}
