--- homework05/src/test/java/me/vladislav/homework/PostgresTestcontainersTest.java
@@ -1,71 +1,0 @@
-package me.vladislav.homework;
-
-import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-import javax.sql.DataSource;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-@Slf4j
-@SpringBootTest
-@Testcontainers
-@ActiveProfiles("test")
-class PostgresTestcontainersTest {
-
-  @Container
-  static PostgreSQLContainer<?> postgres =
-      new PostgreSQLContainer<>("postgres:16-alpine")
-          .withDatabaseName("testdb")
-          .withUsername("postgres")
-          .withPassword("too_strong_password_to_remember")
-          .withInitScript("init.sql");
-
-  @Autowired
-  private DataSource dataSource;
-
-  @DynamicPropertySource
-  static void configureProperties(DynamicPropertyRegistry registry) {
-    registry.add("spring.datasource.url", postgres::getJdbcUrl);
-    registry.add("spring.datasource.username", postgres::getUsername);
-    registry.add("spring.datasource.password", postgres::getPassword);
-    registry.add("spring.datasource.driver-class-name", () -> "org.postgresql.Driver");
-    registry.add("spring.jpa.database-platform", () -> "org.hibernate.dialect.PostgreSQLDialect");
-  }
-
-  @Test
-  void containerStartsAndInitializes() {
-    assertTrue(postgres.isRunning());
-
-    log.info(
-        "PostgreSQL container is running at: "
-            + postgres.getHost()
-            + ":"
-            + postgres.getFirstMappedPort());
-    log.info("JDBC URL: " + postgres.getJdbcUrl());
-  }
-
-  @TestConfiguration
-  static class TestConfig {
-    @Bean
-    public DataSource dataSource() {
-      DriverManagerDataSource dataSource = new DriverManagerDataSource();
-      dataSource.setDriverClassName("org.postgresql.Driver");
-      dataSource.setUrl(postgres.getJdbcUrl());
-      dataSource.setUsername(postgres.getUsername());
-      dataSource.setPassword(postgres.getPassword());
-      return dataSource;
-    }
-  }
-}
