+++ homework05/src/test/java/me/vladislav/homework/app/e2e/BookE2ETest.java
@@ -1,6 +1,7 @@
-package me.vladislav.homework.app;
+package me.vladislav.homework.app.e2e;
 
 import jakarta.annotation.PostConstruct;
+import me.vladislav.homework.app.TestContainersConfig;
 import me.vladislav.homework.app.dto.api.request.BookCreateRequest;
 import me.vladislav.homework.app.dto.api.request.BookPatchRequest;
 import me.vladislav.homework.app.dto.api.request.BookUpdateRequest;
@@ -29,7 +30,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  */
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
-public class BookE2ETest {
+public class BookE2ETest extends TestContainersConfig {
   @Autowired
   private TestRestTemplate restTemplate;
