+++ homework05/src/test/resources/application-test.yml
@@ -4,6 +4,9 @@ spring:
     url: jdbc:postgresql://localhost:5432/homework
     username: CleverOne
     password: compl3xNstrong
+  jpa:
+    hibernate:
+      ddl-auto: create
   application:
     name: homework-test
   flyway:
